requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.22.0
pandas>=1.5.0
matplotlib>=3.5.0
seaborn>=0.12.0
Pillow>=9.0.0
opencv-python>=4.7.0
supervision>=0.18.0
ultralytics>=8.0.0
transformers>=4.36.0
timm>=0.9.5
ftfy>=6.1.0
einops>=0.6.0
pycocotools>=2.0.6
tqdm>=4.65.0
scipy>=1.10.0
scikit-learn>=1.2.0
open_clip_torch>=2.20.0
yacs
nltk
inflect
pyyaml
git+https://github.com/openai/CLIP.git
sentence_transformers